+++ app/javascript/components/atoms/Checkbox.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
-export const Checkbox = ({ id, labelPosition = 'right', checked = false, label, onClick }) => {
+export const Checkbox = ({ labelPosition = 'right', checked = false, label, onClick }) => {
   const renderLabel = (className) => {
     return (
       <label
-        for={`${id}`}
         className={`cursor-pointer ${className}`}
+        onClick={onClick}
       >{label}</label>
     )
   };
@@ -16,16 +16,10 @@ export const Checkbox = ({ id, labelPosition = 'right', checked = false, label, 
       <div className="toggle" onClick={onClick}>
         <input
           checked={checked}
-          id={id}
           type="checkbox"
           className="toggle-checkbox"
         />
-        <label
-          for={id}
-          role="switch"
-          aria-checked="mixed"
-          class="toggle-label"
-        />
+        <label className="toggle-label" />
       </div>
       {label && labelPosition === 'right' ? renderLabel('ml-2') : null}
     </div>